pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hydride-segmentation"
version = "0.5.0"
description = "Toolkit for zirconium hydride segmentation and analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "numpy",
    "matplotlib",
    "opencv-python",
    "scikit-image",
    "scikit-learn",
    "pillow",
    "albumentations",
    "tqdm",
    "scipy",
    "tkinterdnd2",
    "torch",
    "segmentation-models-pytorch",
    "flask",
    "pydantic",
]

[project.optional-dependencies]
gui-qt = ["PySide6"]

[project.scripts]
hydride-gui = "hydride_segmentation.gui:main"
hydride-orientation = "hydride_segmentation.hydride_orientation_analyzer:main"
segmentation-eval = "hydride_segmentation.segmentation_evaluator:main"
hydride-service = "hydride_segmentation.service:main"
hydride-gui-qt = "hydride_segmentation.qt_gui:main"
package-corrections-dataset = "scripts.package_corrections_dataset:main"
